+++ src/hooks/use-state-with-ref.ts
@@ -1,9 +1,17 @@
 import { useEffect, useRef, useState } from 'react';
+import useLocalStorage from './use-local-storage';
 
-const useStateWithRef = <T>(initialValue: T): [T, typeof setValue, typeof searchValueRef] => {
-  const searchValueRef = useRef<T>();
+const useStateWithRef = (storageKey: string): [string, typeof setValue, typeof searchValueRef] => {
+  const searchValueRef = useRef<string>();
+  const [storageSet, storageGet] = useLocalStorage(storageKey);
+  const [value, setValue] = useState<string>(storageGet());
 
-  const [value, setValue] = useState<T>(initialValue);
+  useEffect(() => {
+    const valueRef = searchValueRef.current; // this fixes eslint warning
+    return () => {
+      storageSet(valueRef || '');
+    };
+  }, [searchValueRef, storageSet]);
 
   useEffect(() => {
     searchValueRef.current = value;
